Cleanup
- Cleaned up code in RequestRepository
- Moved table columns to useRequestTableColumns
