+++ src/components/RequestRepository/index.tsx
@@ -1,12 +1,8 @@
-import React, { useContext, useEffect, useMemo, useState } from 'react';
+import React, { useContext, useMemo, useState } from 'react';
 import { CSVLink } from 'react-csv';
 import { Controller, useForm } from 'react-hook-form';
 import { Trans, useTranslation } from 'react-i18next';
-import { useDispatch } from 'react-redux';
-import { Link } from 'react-router-dom';
 import {
-  CellProps,
-  Column,
   Row,
   SortingRule,
   useFilters,
@@ -23,7 +19,6 @@ import {
   Form,
   FormGroup,
   GridContainer,
-  IconError,
   ModalFooter,
   ModalHeading,
   Table
@@ -32,7 +27,6 @@ import classnames from 'classnames';
 import { useFlags } from 'launchdarkly-react-client-sdk';
 import { lowerCase, startCase } from 'lodash';
 
-import UswdsReactLink from 'components/LinkWrapper';
 import Modal from 'components/Modal';
 import PageHeading from 'components/PageHeading';
 import CsvDownloadLink from 'components/shared/CsvDownloadLink';
@@ -40,7 +34,6 @@ import DatePickerFormatted from 'components/shared/DatePickerFormatted';
 import FieldErrorMsg from 'components/shared/FieldErrorMsg';
 import HelpText from 'components/shared/HelpText';
 import Label from 'components/shared/Label';
-import TruncatedText from 'components/shared/TruncatedText';
 import GlobalClientFilter from 'components/TableFilter';
 import TablePageSize from 'components/TablePageSize';
 import TablePagination from 'components/TablePagination';
@@ -50,9 +43,7 @@ import useTableState from 'hooks/useTableState';
 import GetSystemIntakesTableQuery from 'queries/GetSystemIntakesTableQuery';
 import { GetSystemIntakesTable } from 'queries/types/GetSystemIntakesTable';
 import { SystemIntakeState } from 'types/graphql-global-types';
-import { fetchSystemIntakes } from 'types/routines';
 import { cleanCSVData } from 'utils/csv';
-import { formatDateLocal, formatDateUtc } from 'utils/date';
 import globalFilterCellText from 'utils/globalFilterCellText';
 import {
   getColumnSortStatus,
@@ -65,6 +56,7 @@ import { ActiveStateType, TableStateContext } from 'views/TableStateWrapper';
 import csvHeaderMap from './csvHeaderMap';
 import csvPortfolioReportHeaderMap from './csvPortfolioReportHeaderMap';
 import tableMap, { SystemIntakeForTable } from './tableMap';
+import useRequestTableColumns from './useRequestTableColumns';
 
 import './index.scss';
 
@@ -76,10 +68,30 @@ type SystemIntakesData = {
 const RequestRepository = () => {
   const isMobile = useCheckResponsiveScreen('tablet');
   const { t } = useTranslation('governanceReviewTeam');
-  const dispatch = useDispatch();
 
   const flags = useFlags();
 
+  const { itGovAdmin } = useContext(TableStateContext);
+  const [activeTable, setActiveTable] = useState<ActiveStateType>(
+    itGovAdmin.current.activeTableState
+  );
+
+  /* Controls Portfolio Update Report info modal */
+  const [infoModalOpen, setInfoModalOpen] = useState<boolean>(false);
+  /* Controls Configure Portfolio Update Report modal */
+  const [configReportModalOpen, setConfigReportModalOpen] = useState<boolean>(
+    false
+  );
+
+  /* Date range for Portfolio Update Report */
+  const { control, handleSubmit, watch } = useForm<{
+    dateStart: string;
+    dateEnd: string;
+  }>({ resolver: yupResolver(dateRangeSchema) });
+  const dateRangeStart = watch('dateStart');
+  const dateRangeEnd = watch('dateEnd');
+
+  // GQL query to get all system intakes
   const { data: queryData } = useQuery<GetSystemIntakesTable>(
     GetSystemIntakesTableQuery
   );
@@ -107,34 +119,40 @@ const RequestRepository = () => {
     );
   }, [queryData, t]);
 
-  const { itGovAdmin } = useContext(TableStateContext);
-
-  const [activeTable, setActiveTable] = useState<ActiveStateType>(
-    itGovAdmin.current.activeTableState
+  // Returns array of intakes based on active table
+  const data: SystemIntakeForTable[] = useMemo(
+    () => systemIntakes[activeTable],
+    [systemIntakes, activeTable]
   );
 
-  /* Controls Portfolio Update Report info modal */
-  const [infoModalOpen, setInfoModalOpen] = useState<boolean>(false);
-
-  /* Controls Configure Portfolio Update Report modal */
-  const [configReportModalOpen, setConfigReportModalOpen] = useState<boolean>(
-    false
-  );
-
-  /* Date range for Portfolio Update Report */
-  const { control, handleSubmit, watch } = useForm<{
-    dateStart: string;
-    dateEnd: string;
-  }>({
-    resolver: yupResolver(dateRangeSchema),
-    defaultValues: {
-      dateStart: '2023-09-01T05:00:00.000Z',
-      dateEnd: '2023-10-31T05:00:00.000Z'
+  /** Portfolio Update Report data for CSV based on selected date range */
+  const portfolioUpdateReport = useMemo(() => {
+    // If no intakes or selected dates, return empty array
+    if (systemIntakes.closed.length === 0 || !dateRangeStart || !dateRangeEnd) {
+      return [];
     }
-  });
-  const dateRangeStart = watch('dateStart');
-  const dateRangeEnd = watch('dateEnd');
 
+    // Return closed system intakes filtered by selected date range
+    return systemIntakes.closed.filter(({ filterDate }) => {
+      if (!filterDate) return false;
+
+      return filterDate > dateRangeStart && filterDate < dateRangeEnd;
+    });
+  }, [systemIntakes.closed, dateRangeStart, dateRangeEnd]);
+
+  /** Convert selected intakes to CSV format */
+  const convertIntakesToCSV = (
+    intakes: SystemIntakeForTable[]
+  ): SystemIntakeForTable[] => {
+    return intakes.map(intake => {
+      return cleanCSVData(intake);
+    });
+  };
+
+  const csvHeaders = csvHeaderMap(t);
+  const csvPortfolioReportHeaders = csvPortfolioReportHeaderMap(t);
+
+  /** Default page size from local storage */
   const defaultPageSize: number = window.localStorage['request-table-page-size']
     ? Number(window.localStorage['request-table-page-size'])
     : 50;
@@ -156,241 +174,6 @@ const RequestRepository = () => {
     setSortBy(lastSort[nextActiveTable]);
   }
 
-  // Character limit for length of free text (Admin Note, LCID Scope, etc.), any
-  // text longer then this limit will be displayed with a button to allow users
-  // to expand/unexpand the text
-  const freeFormTextCharLimit = 25;
-
-  const submissionDateColumn: Column<SystemIntakeForTable> = {
-    Header: t<string>('intake:fields.submissionDate'),
-    accessor: 'submittedAt',
-    Cell: cell => {
-      if (cell.value) {
-        return formatDateLocal(cell.value, 'MM/dd/yyyy');
-      }
-
-      return t('requestRepository.table.submissionDate.null');
-    }
-  };
-
-  const requestNameColumn: Column<SystemIntakeForTable> = {
-    Header: t<string>('intake:fields.projectName'),
-    accessor: 'requestName',
-    Cell: ({
-      row,
-      value
-    }: CellProps<
-      SystemIntakeForTable,
-      SystemIntakeForTable['requestName']
-    >) => {
-      return (
-        <Link to={`/governance-review-team/${row.original.id}/intake-request`}>
-          {value}
-        </Link>
-      );
-    }
-  };
-
-  const requesterColumn: Column<SystemIntakeForTable> = {
-    Header: t<string>('intake:contactDetails.requester'),
-    accessor: 'requesterNameAndComponent'
-  };
-
-  const adminLeadColumn: Column<SystemIntakeForTable> = {
-    Header: t<string>('intake:fields.adminLead'),
-    accessor: ({ adminLead }) =>
-      adminLead || t<string>('governanceReviewTeam:adminLeads.notAssigned'),
-    Cell: ({
-      value: adminLead
-    }: CellProps<SystemIntakeForTable, SystemIntakeForTable['adminLead']>) => {
-      if (adminLead === t('governanceReviewTeam:adminLeads.notAssigned')) {
-        return (
-          <div className="display-flex flex-align-center">
-            {/* TODO: should probably make this a button that opens up the assign admin
-                lead automatically. Similar to the Dates functionality */}
-            <IconError className="text-secondary margin-right-05" />
-            {adminLead}
-          </div>
-        );
-      }
-      return adminLead;
-    }
-  };
-
-  const grtDateColumn: Column<SystemIntakeForTable> = {
-    Header: t<string>('intake:fields.grtDate'),
-    accessor: 'grtDate',
-    Cell: ({
-      row,
-      value
-    }: CellProps<SystemIntakeForTable, SystemIntakeForTable['grtDate']>) => {
-      if (!value) {
-        return (
-          <UswdsReactLink
-            data-testid="add-grt-date-cta"
-            to={`/governance-review-team/${row.original.id}/dates`}
-          >
-            {t('requestRepository.table.addDate')}
-          </UswdsReactLink>
-        );
-      }
-      return formatDateUtc(value, 'MM/dd/yyyy');
-    }
-  };
-
-  const grbDateColumn: Column<SystemIntakeForTable> = {
-    Header: t<string>('intake:fields.grbDate'),
-    accessor: 'grbDate',
-    Cell: ({
-      row,
-      value
-    }: CellProps<SystemIntakeForTable, SystemIntakeForTable['grbDate']>) => {
-      if (!value) {
-        return (
-          <UswdsReactLink
-            data-testid="add-grb-date-cta"
-            to={`/governance-review-team/${row.original.id}/dates`}
-          >
-            {t('requestRepository.table.addDate')}
-          </UswdsReactLink>
-        );
-      }
-      return formatDateUtc(value, 'MM/dd/yyyy');
-    }
-  };
-
-  const statusColumn: Column<SystemIntakeForTable> = {
-    Header: t<string>('intake:fields.status'),
-    accessor: 'status',
-    Cell: ({
-      row,
-      value
-    }: CellProps<SystemIntakeForTable, SystemIntakeForTable['status']>) => {
-      // If LCID_ISSUED append LCID Scope to status
-      if (value === `LCID: ${row.original.lcid}`) {
-        return (
-          <>
-            {value}
-            <br />
-            <TruncatedText
-              id="lcid-scope"
-              label="less"
-              closeLabel="more"
-              text={`Scope: ${row.original.lcidScope}`}
-              charLimit={freeFormTextCharLimit}
-              className="margin-top-2"
-            />
-          </>
-        );
-      }
-
-      // If any other value just display status
-      return value;
-    }
-  };
-
-  const lcidExpirationDateColumn: Column<SystemIntakeForTable> = {
-    Header: t<string>('intake:fields.lcidExpirationDate'),
-    accessor: 'lcidExpiresAt',
-    Cell: ({ value: lcidExpiresAt }) => {
-      if (lcidExpiresAt) {
-        return formatDateUtc(lcidExpiresAt, 'MM/dd/yyyy');
-      }
-
-      // If no LCID Expiration exists, display 'No LCID Issued'
-      return 'No LCID Issued';
-    }
-  };
-
-  /**
-   * TODO: Fix last admin note column
-   */
-
-  const lastAdminNoteColumn: Column<SystemIntakeForTable> = {
-    Header: t<string>('intake:fields.lastAdminNote'),
-    accessor: 'lastAdminNote',
-    Cell: ({
-      value: lastAdminNote
-    }: {
-      value: SystemIntakeForTable['lastAdminNote'];
-    }) => {
-      if (!lastAdminNote) return 'No Admin Notes';
-
-      return (
-        // Display admin note using truncated text field that
-        // will display note with expandable extra text (if applicable)
-        <>
-          {formatDateLocal(lastAdminNote.createdAt, 'MM/dd/yyyy')}
-
-          <TruncatedText
-            id="last-admin-note"
-            label="less"
-            closeLabel="more"
-            text={lastAdminNote.content}
-            charLimit={freeFormTextCharLimit}
-          />
-        </>
-      );
-    }
-    // TODO: Sort type
-    // sortType: (a: Row<LastAdminNote>, b: Row<LastAdminNote>) =>
-    //   (a.values.lastAdminNote?.createdAt ?? '') >
-    //   (b.values.lastAdminNote?.createdAt ?? '')
-    //     ? 1
-    //     : -1
-  };
-
-  const columns: any = useMemo(() => {
-    if (activeTable === 'open') {
-      return [
-        submissionDateColumn,
-        requestNameColumn,
-        requesterColumn,
-        adminLeadColumn,
-        statusColumn,
-        grtDateColumn,
-        grbDateColumn
-      ];
-    }
-    if (activeTable === 'closed') {
-      return [
-        submissionDateColumn,
-        requestNameColumn,
-        requesterColumn,
-        lcidExpirationDateColumn,
-        statusColumn,
-        lastAdminNoteColumn
-      ];
-    }
-    return [];
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [activeTable, t]);
-
-  // Set active table data
-  const data: SystemIntakeForTable[] = useMemo(
-    () => systemIntakes[activeTable],
-    [systemIntakes, activeTable]
-  );
-
-  /** Portfolio Update Report data for CSV based on selected date range */
-  const portfolioUpdateReport = useMemo(() => {
-    // If no intakes or selected dates, return empty array
-    if (systemIntakes.closed.length === 0 || !dateRangeStart || !dateRangeEnd) {
-      return [];
-    }
-
-    // Return closed system intakes filtered by selected date range
-    return systemIntakes.closed.filter(({ filterDate }) => {
-      if (!filterDate) return false;
-
-      return filterDate > dateRangeStart && filterDate < dateRangeEnd;
-    });
-  }, [systemIntakes.closed, dateRangeStart, dateRangeEnd]);
-
-  useEffect(() => {
-    dispatch(fetchSystemIntakes());
-  }, [dispatch]);
-
   const {
     getTableProps,
     getTableBodyProps,
@@ -411,7 +194,7 @@ const RequestRepository = () => {
     setSortBy
   } = useTable(
     {
-      columns,
+      columns: useRequestTableColumns(activeTable),
       sortTypes: {
         alphanumeric: (rowOne, rowTwo, columnName) => {
           return sortColumnValues(
@@ -436,18 +219,6 @@ const RequestRepository = () => {
   );
 
   rows.map(row => prepareRow(row));
-
-  const csvHeaders = csvHeaderMap(t);
-  const csvPortfolioReportHeaders = csvPortfolioReportHeaderMap(t);
-
-  /** Convert selected intakes to CSV format */
-  const convertIntakesToCSV = (
-    intakes: SystemIntakeForTable[]
-  ): SystemIntakeForTable[] => {
-    return intakes.map(intake => {
-      return cleanCSVData(intake);
-    });
-  };
 
   // Sets persisted table state and stores state on unmount
   useTableState(
